requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.0
Pillow==6.1.0
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
scipy==1.3.0
six==1.12.0
tensorboardX==1.8
torch==1.0.1.post2
torchvision==0.2.1
tqdm==4.32.2
